requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.0
networkx==2.4
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
six==1.15.0
tifffile==2020.7.24
tensorflow>=2.3
keras
numpy<1.19.0
scipy==1.4.1
flask
Flask-Bootstrap

